gai-gen/requirements_tts.txt
torch==2.2.0
torchaudio==2.2.0
transformers==4.36.2
openai==1.6.1
python-dotenv==1.0.0
TTS==0.22.0
uvicorn==0.23.2
ninja==1.11.1.1
fastapi==0.109.2
numpy==1.22
gai-common==0.1.4